App restructuring in preparation for the home page.
