+++ features/second-tier-entitlement/components/NavBarExtension/NavBarExtensionStyles.ts
